added channels to connector
