+++ libraries/Microsoft.Bot.Connector/Channels.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Bot.Connector
+{
+    public class Channels
+    {
+        public const string Facebook = "facebook";
+        public const string Skype = "skype";
+        public const string Msteams = "msteams";
+        public const string Telegram = "telegram";
+        public const string Kik = "kik";
+        public const string Email = "email";
+        public const string Slack = "slack";
+        public const string Groupme = "groupme";
+        public const string Sms = "sms";
+        public const string Emulator = "emulator";
+        public const string Directline = "directline";
+        public const string Webchat = "webchat";
+        public const string Console = "console";
+        public const string Cortana = "cortana";
+    }
+}
